+++ source/packages/@aws-accelerator/constructs/lib/aws-events/security-hub-event-log/test/static-input.ts
@@ -1,0 +1,6 @@
+export abstract class StaticInput {
+  public static readonly newProps = {
+    logGroupName: '/AWSAccelerator-SecurityHub',
+    logGroupArn: 'arn:aws:logs:us-east-1:123456789012:log-group:/*:*',
+  };
+}