+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- fix(logging): refactored Security Hub logging to use EventBridge
+
 ### Fixed
 
 - fix(docs): resolve broken links to appropriate pages